# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 化工厂质检报告系统 API
  description: |
    Node.js + Express 后端接口规范。需登录接口使用 JWT Bearer。
    公开扫码与客户报告接口无需 Bearer（见 Public 分组）。

    本文件覆盖质检核心业务的大部分路径；**仪表盘、部门、备份、报告配图库/样式、
    销售域、企业微信管理 API** 等请同时查阅仓库根目录 `API.md`（§15 起）并对照
    `server/src/routes/` 源码。
  version: 0.3.0
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001
    description: 本地开发
  - url: '{protocol}://{host}'
    description: 部署环境（替换 protocol/host）
    variables:
      protocol:
        default: http
      host:
        default: api.example.com
...
security:
  - bearerAuth: []
paths:
...
  /api/employee-categories:
    get:
      tags: [EmployeeCategories]
      summary: 员工类别列表
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
    post:
      tags: [EmployeeCategories]
      summary: 新建类别
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [nameZh, code]
              properties:
                nameZh: { type: string, maxLength: 64 }
                code:
                  type: string
                  maxLength: 32
                  pattern: '^[a-z][a-z0-9_]*$'
                sortOrder: { type: integer }
                defaultPermissions: { type: object }
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
  /api/employee-categories/{id}:
    parameters:
      - $ref: '#/components/parameters/CategoryId'
    put:
      tags: [EmployeeCategories]
      summary: 更新类别
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nameZh: { type: string }
                sortOrder: { type: integer }
                defaultPermissions: { type: object }
    delete:
      tags: [EmployeeCategories]
      summary: 删除类别
      responses:
        '400':
          content:
            application/json:
              examples:
                builtin:
                  value:
                    error: CANNOT_DELETE_BUILTIN
                inUse:
                  value:
                    error: CATEGORY_IN_USE
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录接口返回的 token
  parameters:
...
    CategoryId:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...